2x speedup on minimizeConflicts and added additional info panel for conflicts to displayed schedules.
To achieve a 2x speedup on minimizeConflicts, getOptimizedVariable was made much more efficient. Previously, it had run vanilla weightedConflicts for each hypothetical variable to guage its efficiency. This was inefficient because only one variable is changed, so most constraints remain unaffected by the tested values. To combat this, a new attribute is stored in CSP: constraintsByVariable. It stores which constraints are affected by each variable and which aren't. getOptimizedVariable can then store a baseConflicts as the conflicts caused by constraints out of the reach of the variable-to-optimize. This can then be added to the weightedConflicts of only the constraints affected by the variable within the loop. Thus, getOptimizedVariable's loop only needs to check about 100 constraints in its loop opposed to 2000 before.