Conflicting variables stored and updated.
Previously, conflicting variables were computed at every minimizeConflicts step, and one was randomly chosen to be optimized. This was inefficient because certain variables would already be where they minimized conflicts, but they would be called to get optimized again: the program explored dead ends. Now, a list of conflicting variables is calculated at the beginning of minimizeConflicts and recalculated whenever a variable changes, because changes can induce new unforeseen conflicts. When a variable is optimized but doesn't change, however, it is removed from the list of conflictingVariables so that it isn't called again until a different variable has changed position and made it possible that an optimization exists.
By storing and dynamically updating a list of conflicting variables, the need to check for plateaus is removed. Instead, if the length of the list of conflicting variables ever goes to zero, no variables can be improved, and the program can stop. Thus, a dynamic plateau indication is embedded within the list of conflicting variables.