Added milestone functionality to discard unpromising schedules.
Before starting to optimize, an array of milestones can be calculated by taking how many conflicts the xth percentile of schedules has after a given number of steps. Schedules whose optimizations don't meet the benchmarks are terminated immediately. The goal is that few good schedules are toosed out but many bad schedules are removed from the search. This change came from the observation that almost half of fully optimized schedules end with over a million weighted conflicts and the desire to toss them early.